+++ CHANGELOG.rst
@@ -20,6 +20,10 @@ Changelog
   methods to ECDSA keys.
 * Switched back to the older callback model on Python 3.5 in order to mitigate
   the locking callback problem with OpenSSL <1.1.0.
+* :class:`~cryptography.x509.CertificateBuilder` now also accepts aware
+  datetimes for 
+  :meth:`~cryptography.x509.CertificateBuilder.not_valid_before` and
+  :meth:`~cryptography.x509.CertificateBuilder.not_valid_after`
 
 
 1.4 - 2016-06-04
@@ -51,6 +55,8 @@ Changelog
 
 * Added two new OpenSSL functions to the bindings to support an upcoming
   ``pyOpenSSL`` release.
+* Support for OpenSSL 0.9.8 has been removed. Users on older version of OpenSSL
+  will need to upgrade.
 
 1.3.2 - 2016-05-04
 ~~~~~~~~~~~~~~~~~~